remove DarwinDataset
