+++ darwin/dataset/local_dataset.py
@@ -1,8 +1,11 @@
 import json
+import multiprocessing as mp
 from pathlib import Path
-from typing import List, Optional
+from typing import Collection, List, Optional
 
-from darwin.dataset.utils import get_classes, get_release_path
+import numpy as np
+
+from darwin.dataset.utils import get_classes, get_release_path, load_pil_image
 from darwin.utils import SUPPORTED_IMAGE_EXTENSIONS
 
 
@@ -178,6 +181,97 @@ class LocalDataset(object):
             "width": data["image"]["width"],
             "annotations": annotations,
         }
+
+    def measure_mean_std(self, multi_threaded: bool = True):
+        """Computes mean and std of train images, given the train loader
+
+        Parameters
+        ----------
+        multi_threaded : bool
+            Uses multiprocessing to download the dataset in parallel.
+
+        Returns
+        -------
+        mean : ndarray[double]
+            Mean value (for each channel) of all pixels of the images in the input folder
+        std : ndarray[double]
+            Standard deviation (for each channel) of all pixels of the images in the input folder
+        """
+        if multi_threaded:
+            # Set up a pool of workers
+            with mp.Pool(mp.cpu_count()) as pool:
+                # Online mean
+                results = pool.map(self._return_mean, self.images_path)
+                mean = np.sum(np.array(results), axis=0) / len(self.images_path)
+                # Online image_classification deviation
+                results = pool.starmap(
+                    self._return_std, [[item, mean] for item in self.images_path]
+                )
+                std_sum = np.sum(np.array([item[0] for item in results]), axis=0)
+                total_pixel_count = np.sum(np.array([item[1] for item in results]))
+                std = np.sqrt(std_sum / total_pixel_count)
+                # Shut down the pool
+                pool.close()
+                pool.join()
+            return mean, std
+        else:
+            # Online mean
+            results = [self._return_mean(f) for f in self.images_path]
+            mean = np.sum(np.array(results), axis=0) / len(self.images_path)
+            # Online image_classification deviation
+            results = [self._return_std(f, mean) for f in self.images_path]
+            std_sum = np.sum(np.array([item[0] for item in results]), axis=0)
+            total_pixel_count = np.sum(np.array([item[1] for item in results]))
+            std = np.sqrt(std_sum / total_pixel_count)
+            return mean, std
+
+    def measure_weights(self, **kwargs):
+        """Computes the class balancing weights (not the frequencies!!) given the train loader
+
+        Returns
+        -------
+        class_weights : ndarray[double]
+            Weight for each class in the train set (one for each class)
+        """
+        raise NotImplementedError("Base class Dataset does not have an implementation for this")
+
+    @staticmethod
+    def _compute_weights(labels: Collection):
+        """Given an array of labels computes the weights normalized
+
+        Parameters
+        ----------
+        labels : ndarray[int]
+            Array of labels
+
+        Returns
+        -------
+        ndarray[float]
+            Array of weights (one for each unique class) which are the inverse of their frequency
+        """
+        class_support = np.unique(labels, return_counts=True)[1]
+        class_frequencies = class_support / len(labels)
+        # Class weights are the inverse of the class frequencies
+        class_weights = 1 / class_frequencies
+        # Normalize vector to sum up to 1.0 (in case the Loss function does not do it)
+        class_weights /= class_weights.sum()
+        return class_weights
+
+    # Loads an image with OpenCV and returns the channel wise means of the image.
+    @staticmethod
+    def _return_mean(image_path):
+        img = np.array(load_pil_image(image_path))
+        mean = np.array([np.mean(img[:, :, 0]), np.mean(img[:, :, 1]), np.mean(img[:, :, 2])])
+        return mean / 255.0
+
+    # Loads an image with OpenCV and returns the channel wise std of the image.
+    @staticmethod
+    def _return_std(image_path, mean):
+        img = np.array(load_pil_image(image_path)) / 255.0
+        m2 = np.square(
+            np.array([img[:, :, 0] - mean[0], img[:, :, 1] - mean[1], img[:, :, 2] - mean[2]])
+        )
+        return np.sum(np.sum(m2, axis=1), 1), m2.size / 3.0
 
     def __getitem__(self, index: int):
         return self.parse_json(index)
