+++ darwin/dataset/utils.py
@@ -9,6 +9,7 @@ from pathlib import Path
 from typing import Generator, Iterable, List, Optional, Union
 
 import numpy as np
+from PIL import Image
 from tqdm import tqdm
 
 from darwin.exceptions import NotFound
@@ -682,3 +683,40 @@ def get_annotations(
             with annotation_path.open() as f:
                 record = json.load(f)
             yield record
+
+
+def load_pil_image(path: Path):
+    """
+    Loads a PIL image and converts it into RGB.
+
+    Parameters
+    ----------
+    path: Path
+        Path to the image file
+
+    Returns
+    -------
+    PIL Image
+        Values between 0 and 255
+    """
+    pic = Image.open(path)
+    if pic.mode == "RGB":
+        pass
+    elif pic.mode in ("CMYK", "RGBA", "P"):
+        pic = pic.convert("RGB")
+    elif pic.mode == "I":
+        img = (np.divide(np.array(pic, np.int32), 2 ** 16 - 1) * 255).astype(np.uint8)
+        pic = Image.fromarray(np.stack((img, img, img), axis=2))
+    elif pic.mode == "I;16":
+        img = (np.divide(np.array(pic, np.int16), 2 ** 8 - 1) * 255).astype(np.uint8)
+        pic = Image.fromarray(np.stack((img, img, img), axis=2))
+    elif pic.mode == "L":
+        img = np.array(pic).astype(np.uint8)
+        pic = Image.fromarray(np.stack((img, img, img), axis=2))
+    else:
+        raise TypeError(f"unsupported image type {pic.mode}")
+    return pic
+
+
+def _is_pil_image(img):
+    return isinstance(img, Image.Image)
